requirements.txt
# Self-Improving LLM - Agent Dependencies
# Subset of TuneLLM deps needed for training, evaluation, and infra management

# PyTorch (CPU for local, GPU on remote instances)
torch>=2.1.0

# Transformers & LLM
transformers>=4.36.0
datasets>=2.16.0
tokenizers>=0.15.0
accelerate>=0.25.0
safetensors>=0.4.1

# Parameter-Efficient Fine-Tuning
peft>=0.7.1

# RLHF
trl>=0.7.6

# Quantization
bitsandbytes>=0.41.3

# Evaluation & Metrics
evaluate>=0.4.1
rouge-score>=0.1.2
nltk>=3.8.1
scikit-learn>=1.3.2

# Configuration
pyyaml>=6.0.1

# Monitoring & Logging
wandb>=0.16.2
rich>=13.7.0

# Utilities
numpy>=1.26.3
pandas>=2.1.4
tqdm>=4.66.1

# HTTP Client (for Vast.ai, SSH infra, Paperclip API)
httpx>=0.26.0

# System Monitoring
psutil>=5.9.8
